propper static member function declaration
testing static function in encoder unit test

+++ Libraries/Encoder/Encoder.h
@@ -5,18 +5,19 @@
 
 class Encoder
 {
+
   public:
     Encoder(int8_t PinA, int8_t PinB, volatile int32_t *Ticks, volatile int32_t *Velocity, boolean Flip);
     void tick();
     void compute();
-	static void staticCompute(uint8_t objectPointer);
+	static void staticCompute(int objectPointer);
     int32_t getPosition();
     void setPosition(const int32_t p);
     void setLowPinA();
     void setHighPinA();
     void setLowPinB();
     void setHighPinB();
-    
+
   private:
     volatile int32_t *position; 
     volatile int32_t *velocity;
